Little code improvements to config
